Add login functionality using JWT for user authentication

+++ src/main/java/com/example/job/portal/Repository/EmployerRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface EmployerRepo extends JpaRepository<Employer, Integer> {
+public interface EmployerRepo extends JpaRepository<Employer, Long> {
     Optional<Employer> findByEmail(String email);
 }
