+++ src/main/java/com/example/job/portal/Repository/SeekerRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface SeekerRepo extends JpaRepository<Seeker, Integer> {
+public interface SeekerRepo extends JpaRepository<Seeker, Long> {
     Optional<Seeker> findByEmail(String email);
 }
